Update README.md
In this project, you will implement a Polynomial ADT using Linked Lists. You will implement the Multiplication, Addition, and Subtraction operations of polynomials.
Your program should be able to read a file of polynomials called equations.txt and store each equation in a doubly-linked list. Once the file is read and equations are loaded into the linked lists, the user should be able to perform mathematical operations on the polynomials (addition, subtraction, and multiplication). Once the user select an operation through a menu, the output should be displayed and then the menu should show again. Another option for the user is to store the results of all operations in a file called results.txt. Your application should show an appropriate screen with menu options to inform the user of the available operations to let her/him to choose from.
Example of input file:
2x^7+10x^5-10x^3+2x+1
-15x^7-10x^5+90x^2-2
2x^2+x-1

Upon user selection, the application should show the output of the operation performed (i.e., Addition, Subtraction, or Multiplication). The user then should have an option to enter a value of the variable term in the equation to substitute.